+++ js/util.js
@@ -1,4 +1,3 @@
-const RERENDER_DELAY = 500;
 const addClass = (element, classToAdd) => {
   const className = element.classList.add(classToAdd);
   return className;
@@ -9,14 +8,6 @@ const removeClass = (element, classToRemove) => {
   return classRemove;
 };
 
-const debounce = (callback, timeoutDelay = RERENDER_DELAY) => {
-  let timeoutId;
-  return (...rest) => {
-    clearTimeout(timeoutId);
-    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
-  };
-};
-
 const isEscapeKey = (evt) => evt.key === 'Escape';
 
-export {addClass, removeClass, isEscapeKey, debounce};
+export {addClass, removeClass, isEscapeKey};
